fix: resolve RawMaterial type mismatch and compilation errors in inventory module

+++ module-inventory/src/main/java/com/storyline/erp/inventory/dto/RawMaterialDto.java
@@ -4,12 +4,14 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record RawMaterialDto(
         Long id,
         @NotBlank(message = "SKU is required") String sku,
         @NotBlank(message = "Name is required") String name,
         @NotBlank(message = "Unit of measure is required") String unitOfMeasure,
-        @PositiveOrZero(message = "Current stock cannot be negative") Double currentStock,
-        @PositiveOrZero(message = "Minimum stock cannot be negative") Double minimumStock,
-        @PositiveOrZero(message = "Unit cost cannot be negative") Double unitCost
+        @PositiveOrZero(message = "Current stock cannot be negative") BigDecimal currentStock,
+        @PositiveOrZero(message = "Minimum stock cannot be negative") BigDecimal minimumStock,
+        @PositiveOrZero(message = "Unit cost cannot be negative") BigDecimal unitCost
 ) {}
